+++ service/src/main/java/com/jacinto/model/Cliente.java
@@ -3,17 +3,17 @@ package com.jacinto.model;
 public class Cliente {
 	
 	public Integer id;
-	public Integer limite;
-	public Integer saldo;
+	public Long limite;
+	public Long saldo;
 	
-	public Cliente(Integer id, Integer limite, Integer saldo) {
+	public Cliente(Integer id, Long limite, Long saldo) {
 		super();
 		this.id = id;
 		this.limite = limite;
 		this.saldo = saldo;
 	}
 	
-	public Cliente(Integer limite, Integer saldo) {
+	public Cliente(Long limite, Long saldo) {
 		this(null, limite, saldo);
 	}
